polymarket_decoder/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.0.0